requirements.txt
selenium>=4.15.0
playwright>=1.40.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
tkcalendar>=1.6.1